overlord/configstate: prepare for experimental flags
This patch moves the list of known experimental flags to the new file
where we will also soon handle exporting them to C parts of snapd.
